+++ Changelog.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project follows to [Ragnarök Versioning Convention](https://shor.cz/ragnarok_versioning_convention).
 
+## [UNRELEASED] Red Core Version 0.5 Changelog
+
+### Highlight
+
+Welcome to Red Core 0.5 a breaking release.
+
+
+### Changed
+
+- Changed GroupId from `io.redstudioragnarok` to `dev.redstudio`
+
+---
+
 ## Red Core Version 0.4 Changelog - 2023-09-07
 
 ### Highlight